Began porting .../lib/App/SCS.pm to lib/app/ombre.rb
One of the intentions I have with this port is to be very thorough in my
writing and documentation for it; when I ran into the differences
between the way Ruby handles attributes and Perl's Moo handles
attributes, I had to break them down a little bit and add in my own bit
of laziness in the code.

When I did that, I felt like the code was bare, undescriptive again; so
I pulled up the example docs for YARD and am now gonna document, if only
lightly, all my methods and everything I can. It'll help me remember
what I've figured out in the Perl code.